refactor: change resource itens and folder
