+++ src/main/java/com/example/jvspace/controllers/exceptions/ResourceExceptionsHandler.java
@@ -1,6 +1,6 @@
 package com.example.jvspace.controllers.exceptions;
 
-import com.example.jvspace.services.exceptions.UserNotFoundException;
+import com.example.jvspace.services.exceptions.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -10,10 +10,10 @@ import javax.servlet.http.HttpServletRequest;
 import java.time.Instant;
 
 @ControllerAdvice
-public class UserExceptionsHandler {
+public class ResourceExceptionsHandler {
 
-    @ExceptionHandler(UserNotFoundException.class)
-    public ResponseEntity<StandardError> resourceNotFound(UserNotFoundException e, HttpServletRequest request){
+    @ExceptionHandler(ResourceNotFoundException.class)
+    public ResponseEntity<StandardError> resourceNotFound(ResourceNotFoundException e, HttpServletRequest request){
         String message = "Resource not found";
         HttpStatus status = HttpStatus.NO_CONTENT;
         StandardError err = new StandardError(Instant.now(), status.value(), message, e.getMessage(), request.getRequestURI());
