+++ src/main/java/com/example/jvspace/entities/dto/PostDTO.java
@@ -1,4 +1,4 @@
-package com.example.jvspace.dto;
+package com.example.jvspace.entities.dto;
 
 import com.example.jvspace.entities.Post;
 import com.fasterxml.jackson.annotation.JsonFormat;
