+++ src/main/java/com/example/jvspace/services/PostService.java
@@ -2,6 +2,8 @@ package com.example.jvspace.services;
 
 import com.example.jvspace.entities.Post;
 import com.example.jvspace.repositories.PostRepository;
+import com.example.jvspace.services.exceptions.ResourceNotFoundException;
+import com.example.jvspace.utils.ValidationTools;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,9 +12,9 @@ import java.util.Optional;
 
 @Service
 public class PostService {
-
     @Autowired
     private PostRepository postRepository;
+
 
     public List<Post> findAll(){
         return postRepository.findAll();
@@ -20,7 +22,7 @@ public class PostService {
 
     public Post findById(String id){
         Optional<Post> postObject = postRepository.findById(id);
-        return postObject.orElseThrow(() -> new RuntimeException());
+        return postObject.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public Post insert(Post postObject){
@@ -30,12 +32,12 @@ public class PostService {
     public Post update(String id, Post post){
         try{
             Post updateEntity = findById(id);
-            updateData(updateEntity, post);
+            ValidationTools.updatePostVerify(updateEntity, post);
             return postRepository.save(updateEntity);
         }
         catch (RuntimeException e){
             e.printStackTrace();
-            throw new IllegalArgumentException();
+            throw new ResourceNotFoundException(e.getMessage());
         }
     }
 
@@ -43,13 +45,9 @@ public class PostService {
         try{
             postRepository.deleteById(id);
         }
-        catch (IllegalArgumentException e){
-            throw new IllegalArgumentException();
+        catch (RuntimeException e){
+            throw new ResourceNotFoundException(id);
         }
     }
 
-    private void updateData(Post entity, Post post) {
-        entity.setTitle(post.getTitle());
-        entity.setBody(post.getBody());
-    }
 }
