+++ src/main/java/com/example/jvspace/services/exceptions/ResourceNotFoundException.java
@@ -1,9 +1,9 @@
 package com.example.jvspace.services.exceptions;
 
-public class UserNotFoundException extends RuntimeException{
+public class ResourceNotFoundException extends RuntimeException{
     private static final Long seriaVersionUID = 1L;
 
-    public UserNotFoundException(Object id){
+    public ResourceNotFoundException(Object id){
         super("Resource not found. ID: " + id);
     }
 }
